+++ Cargo.toml
@@ -103,6 +103,7 @@ members = [
     "rpc-client",
     "rpc-client-api",
     "rpc-client-nonce-utils",
+    "rpc-response",
     "rpc-test",
     "runtime",
     "runtime-transaction",
@@ -597,6 +598,7 @@ solana-rpc = { path = "rpc", version = "=2.2.0" }
 solana-rpc-client = { path = "rpc-client", version = "=2.2.0", default-features = false }
 solana-rpc-client-api = { path = "rpc-client-api", version = "=2.2.0" }
 solana-rpc-client-nonce-utils = { path = "rpc-client-nonce-utils", version = "=2.2.0" }
+solana-rpc-response = { path = "rpc-response", version = "=2.2.0" }
 solana-runtime = { path = "runtime", version = "=2.2.0" }
 solana-runtime-transaction = { path = "runtime-transaction", version = "=2.2.0" }
 solana-sbpf = "=0.9.0"
