+++ rpc-client-api/Cargo.toml
@@ -28,6 +28,7 @@ solana-fee-calculator = { workspace = true, features = ["serde"] }
 solana-inflation = { workspace = true }
 solana-inline-spl = { workspace = true }
 solana-pubkey = { workspace = true }
+solana-rpc-response = { workspace = true }
 solana-signer = { workspace = true }
 solana-transaction-error = { workspace = true }
 solana-transaction-status-client-types = { workspace = true }
